Implement basic command handling
This commit implements a basic command handler and a simple ping
command for validation. The command loader recursively iterates
through each command file to validate and load all commands.

A `deploy-commands` script has been defined to register commands
as they only need to be registered once, and there is a daily limit
on command creations, so this should only be run when changes have
been made to the slash command definitions (not execution/handling).

This commit also refactors various pieces of code to improve code
readability and maintainability.