Add purge command and modal boilerplate
This commit introduces the foundations of the purge command. This
includes some basic command processing logic and a modal UI
component for input.

Some refactoring was needed in order to efficiently allow for
command lookups based on modal id, since discord.js doesn't seem
to naturally associate modal submission events with the commands
that produced them.